Match structure of victory-chart
