--- meta.js
@@ -1,7 +1,0 @@
-"use strict";
-
-// Meta configuration for webpack, etc.
-module.exports = {
-  FILE_NAME: "victory-examples",
-  LIB_NAME: "VictoryExamples"
-};
